Move Selection from molgenis-api-data to molgenis-api
