+++ molgenis-api-data/pom.xml
@@ -42,24 +42,6 @@
           </execution>
         </executions>
       </plugin>
-      <plugin>
-        <groupId>com.helger.maven</groupId>
-        <artifactId>ph-javacc-maven-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>jjc1</id>
-            <phase>generate-sources</phase>
-            <goals>
-              <goal>javacc</goal>
-            </goals>
-            <configuration>
-              <jdkVersion>1.8</jdkVersion>
-              <sourceDirectory>src/main/resources</sourceDirectory>
-              <outputDirectory>${project.build.directory}/generated-sources/javacc</outputDirectory>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
     </plugins>
   </build>
