+++ molgenis-api-data/src/main/java/org/molgenis/api/data/v3/DataApiV3Config.java
@@ -1,5 +1,6 @@
 package org.molgenis.api.data.v3;
 
+import org.molgenis.api.convert.SelectionConverter;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.format.FormatterRegistry;
