+++ molgenis-api-data/src/main/java/org/molgenis/api/data/v3/EntityMapper.java
@@ -1,5 +1,6 @@
 package org.molgenis.api.data.v3;
 
+import org.molgenis.api.model.Selection;
 import org.molgenis.data.Entity;
 
 public interface EntityMapper {
