+++ molgenis-api-data/src/main/java/org/molgenis/api/data/v3/EntityMapperImpl.java
@@ -10,6 +10,7 @@ import java.util.Map;
 import javax.annotation.CheckForNull;
 import javax.annotation.Nullable;
 import org.molgenis.api.data.v3.EntityCollection.Page;
+import org.molgenis.api.model.Selection;
 import org.molgenis.api.model.response.LinksResponse;
 import org.molgenis.api.model.response.PageResponse;
 import org.molgenis.data.Entity;
