+++ molgenis-api-data/src/test/java/org/molgenis/api/data/v3/EntityControllerTest.java
@@ -9,6 +9,7 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Map;
 import org.mockito.Mock;
+import org.molgenis.api.model.Selection;
 import org.molgenis.data.Entity;
 import org.molgenis.test.AbstractMockitoTest;
 import org.springframework.http.ResponseEntity;
