+++ molgenis-api-data/src/test/java/org/molgenis/api/data/v3/EntityMapperImplTest.java
@@ -7,8 +7,8 @@ import static java.util.Collections.singletonMap;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.molgenis.api.data.v3.Selection.EMPTY_SELECTION;
-import static org.molgenis.api.data.v3.Selection.FULL_SELECTION;
+import static org.molgenis.api.model.Selection.EMPTY_SELECTION;
+import static org.molgenis.api.model.Selection.FULL_SELECTION;
 import static org.molgenis.data.meta.AttributeType.BOOL;
 import static org.molgenis.data.meta.AttributeType.CATEGORICAL;
 import static org.molgenis.data.meta.AttributeType.CATEGORICAL_MREF;
@@ -39,6 +39,7 @@ import java.util.Iterator;
 import java.util.List;
 import org.mockito.quality.Strictness;
 import org.molgenis.api.data.v3.EntityCollection.Page;
+import org.molgenis.api.model.Selection;
 import org.molgenis.api.model.response.LinksResponse;
 import org.molgenis.api.model.response.PageResponse;
 import org.molgenis.data.Entity;
